requirements.txt
beautifulsoup4==4.6.0
requests==2.19.1
discord.py==0.16.12
dialogflow==0.5.1